+++ tests/docker/php/entrypoint.sh
@@ -1,9 +1,0 @@
-#!/bin/sh
-
-set -eu
-
-flock tests/runtime/composer-install.lock composer install --prefer-dist --no-interaction
-
-php --version
-set -x
-exec "$@"